+++ muan/teleop/button.h
@@ -58,7 +58,7 @@ class PovRange : public Button {
  public:
   friend class muan::teleop::Joystick;
 
-  PovRange(Joystick* joystick, uint32_t button, int minimum, int maximum);
+  PovRange(Joystick* joystick, uint32_t button, double minimum, double maximum);
 
   void Update();
